requirements.txt
gradio
modelscope_studio
dashscope
google-genai
nltk
pypdf
sentence-transformers
numpy
